refactor(ManageVolume): use new VolumeDetail component
