+++ src/components/common/VolumeDetail/hook.ts
@@ -36,7 +36,7 @@ export function useVolumeDetail({
   const [volume] = entities || []
 
   const manager = useVolumeManager()
-  const { next, stop } = useCheckoutNotification({})
+  const { next, stop, noti } = useCheckoutNotification({})
 
   const handleCopyHash = useCopyToClipboardAndNotify(volume?.id || '')
 
@@ -61,10 +61,22 @@ export function useVolumeDetail({
 
       await router.replace('/')
     } catch (e) {
+      console.error(e)
+
+      const text = (e as Error).message
+      const cause = (e as Error)?.cause as string | Error | undefined
+      const detail = typeof cause === 'string' ? cause : cause?.message
+
+      noti?.add({
+        variant: 'error',
+        title: 'Error',
+        text,
+        detail,
+      })
     } finally {
       await stop()
     }
-  }, [dispatch, manager, volume, next, router, stop])
+  }, [manager, volume, dispatch, router, next, noti, stop])
 
   const handleDownload = useCallback(async () => {
     if (!manager) throw Err.ConnectYourWallet
